Bug fix for APEX 5.1
When „Page Items to Submit“ for the region is not set, the APEX 5.1
version of apex.server.plugin has problems with the delivered array
with one empty string element. jQuery reports then a syntax error and
the graph stops loading. This was running under APEX 5.0. By reading
the documentation, I can clearly see, this is more my fault then the
one of APEX - I should have deliver false instead of an array with one
empty string element.